comment delete

+++ register_login/includes/deleteComments.php
@@ -1,10 +1,21 @@
 <?php
-function deleteComments($conn){
+    $serverName = "localhost";
+    $dBUserName = "root";
+    $dBPassword = "";
+    $dBName = "ita212";
+    
+    $conn = mysqli_connect($serverName, $dBUserName, $dBPassword, $dBName);
+    
+    if (!$conn){
+        die("Connection failed". mysqli_connect_error());
+    }
+    
+
     if (isset($_POST['commentDelete'])){
         $cid = $_POST['cid'];
         $sql = "DELETE FROM comments WHERE `comments`.`cid` = $cid";
         // $sql = "DELETE FROM comments WHERE cid='$cid'";
         $result = $conn->query($sql);
-        header("location: ../book.php");
+        echo "TEST";
+        header("Location: ../book.php");
     }  
-}
